+++ src/async.h
@@ -26,4 +26,4 @@ void async_dir_load(Dir *dir, bool dircounts);
 void async_preview_check(Preview *pv);
 
 // Reloads preview of the file at `path` with `nrow` lines from disk.
-void async_preview_load(Preview *pv, uint16_t nrow);
+void async_preview_load(Preview *pv, uint32_t nrow);
